backend/requirements.txt
fastapi
uvicorn
pydantic
numpy
pandas
scikit-learn
python-multipart
python-dotenv
